+++ client/app/modules/admin/admin.module.js
@@ -8,13 +8,14 @@ import ScheduleManagerModule from '../../modules/schedulemanager/schedulemanager
 import UserManagerModule from '../../modules/usermanager/usermanager.module';
 import ProductManagerModule from '../../modules/productmanager/productmanager.module';
 import ClassManagerModule from '../../modules/classmanager/classmanager.module';
+import WorkshopManagerModule from '../../modules/workshopmanager/workshopmanager.module';
 import LocationManagerModule from '../../modules/locationmanager/locationmanager.module';
 import { FileService } from '../../services/file.service';
 
 export default angular.module('shyApp.admin', [
   ngRoute, AnnouncementManagerModule, UibTabsDirective,
   ScheduleManagerModule, UserManagerModule, ProductManagerModule,
-  ClassManagerModule, LocationManagerModule
+  ClassManagerModule, WorkshopManagerModule, LocationManagerModule
 ])
   .service('FileService', FileService)
   .config(routes)