+++ server/api/workshop/workshop.controller.js
@@ -3,26 +3,22 @@ import db from '../../utils/db';
 // Returns JSON array of upcoming workshops in chronological order
 export async function index(req, res) {
   const { rows } = await db.query('SELECT * FROM workshop_sections;', []);
-  return res.status(200).send(rows);
+  return res.status(200).send(rows[0].workshops);
 }
 
 // Ingests JSON representation of workshop including sections (broken out from upsert() for testing)
 export async function upsertWorkshop(workshop) {
-  const { rows } = await db.query('SELECT workshop_upsert($1::json)', [workshop]);
-  const { _id, imageId } = rows[0];
-  const confirmation = {
-    id: _id,
-    imageId
-  };
-  return confirmation;
+  const jsonWorkshop = JSON.stringify(workshop);
+  const { rows } = await db.query('SELECT * FROM workshop_upsert_json($1::json)', [jsonWorkshop]);
+  return rows[0];
 }
 
 // Updates or creates workshop (admin-only)
 export async function upsert(req, res) {
   // Check if workshop was sent
-  if(!req.body.workshop) throw new Error('Did not receive JSON from client.');
-  const _id = await upsertWorkshop(req.body.workshop);
-  return res.status(200).send({ id: _id });
+  if(!req.body) throw new Error('Did not receive JSON from client.');
+  const response = await upsertWorkshop(req.body);
+  return res.status(200).send(response);
 }
 
 // Deletes workshop and cascades to sections (admin-only)